Moves Webhook & Role Helper To Discord File
Moves the webhook helper and the role assignment helper to the discord
file to gather all discord helpers in one location.
